add args dont_explore_names dont_explore_paths ignore_zero_size and use uwalk

+++ finddupes_in_dir.py
@@ -4,6 +4,7 @@ import os
 import argparse
 import filecmp
 import hashlib
+from uwalk import uwalk
 
 description = """
 Any files found in '--dir' that are dupes of files found in '--d' directories 
@@ -14,6 +15,9 @@ parser.add_argument('--d', required=True, action="append", help="where to look f
 parser.add_argument('--dir', required=True, help='where to look for dupes')
 parser.add_argument('--dryrun', action='store_true', help='just look for dupes, does not delete files')
 parser.add_argument('--force', action='store_true', help='do not ask for confirmation to proceed')
+parser.add_argument('-n', '--dont_explore_names', action='append', help='directory names not to explore e.g. .git', default=[])
+parser.add_argument('-p', '--dont_explore_paths', action='append', help='directory paths not to explore e.g. .git', default=[])
+parser.add_argument('--ignore_zero_size', action='store_true', help='ignore any file with zero size')
 args = parser.parse_args()
 print("looking for dupes, deleting from %s"%args.dir)
 
@@ -22,6 +26,8 @@ if not args.dryrun and not args.force:
     user_continue = input('WARNING this is not a dry run. Any dupes found will be deleted. Enter 1 to quit now.')
     if user_continue == '1':
         exit()
+else:
+    print("DRY RUN")
 
 
 def get_file_md5(file_path):
@@ -41,7 +47,8 @@ class DupeFinder():
 
     def add_file(self, file_path):
         size = os.path.getsize(file_path)
-        self.found.setdefault(size, []).append(file_path)
+        if not args.ignore_zero_size or size != 0:
+            self.found.setdefault(size, []).append(file_path)
 
     def prep_size(self, size):
         if size in self.found:
@@ -75,22 +82,32 @@ def dir_file_matched(dir_file_path, dupe_finder):
 # process files from the '--d' args and create lookup structure
 dupe_finder = DupeFinder()
 for path in args.d:
-    for dirpath, dirnames, files in os.walk(path):
+    for dirpath, dirnames, files in uwalk(path, 
+            dont_explore_names=args.dont_explore_names,
+            dont_explore_paths=args.dont_explore_paths):
         for name in files:
             file_path = os.path.join(dirpath, name)
             dupe_finder.add_file(file_path)
 
 
 # process files from the '--dir' arg and delete any that exist in lookup structure
-for dirpath, dirnames, files in os.walk(args.dir):
+for dirpath, dirnames, files in uwalk(args.dir, 
+        dont_explore_names=args.dont_explore_names,
+        dont_explore_paths=args.dont_explore_paths):
     for name in files:
         dir_file_path = os.path.join(dirpath, name)
-        match = dir_file_matched(dir_file_path, dupe_finder)
-        if match:
-            size = os.path.getsize(dir_file_path)
-            print(name, size, '"%s"'%match, '"%s"'%dirpath, "will delete %s" % dir_file_path)
-            try:
-                if not args.dryrun:
-                    os.remove(os.path.join(dirpath, name))
-            except EnvironmentError as e:
-                print("ERROR deleting %s"%os.path.join(dirpath, name))
+        size = os.path.getsize(dir_file_path)
+        if not args.ignore_zero_size or size != 0:
+            match = dir_file_matched(dir_file_path, dupe_finder)
+            if match:
+                print("============================")
+                print("Name: %s, %s" % (name, size))
+                print('             ', '"%s"' % match)
+                print('will delete :', '"%s"' % dir_file_path)
+                # print(name, size, '"%s"'%match, '"%s"'%dirpath, "will delete %s" % dir_file_path)
+                try:
+                    if not args.dryrun:
+                        os.remove(os.path.join(dirpath, name))
+                except EnvironmentError as e:
+                    print("ERROR deleting %s"%os.path.join(dirpath, name))
+
